switch to floating point numbers

+++ 1b_arithmatic/calc.c
@@ -1,16 +1,18 @@
 #include<stdio.h>
 
+// Using %g format specifier to truncate floating numbers
+
 void main() {
-    int operand1, operand2;
+    float operand1, operand2;
     printf("Enter your two operands: ");
-    scanf("%d%d", &operand1, &operand2);
+    scanf("%g%g", &operand1, &operand2);
 
     int operator_choice;
     printf("Enter the operator choice:\n");
     printf("1) Add\n2) Subtract\n3) Multiply\n4) Divide\n\n");
     scanf("%d", &operator_choice);
 
-    int result;
+    float result;
     switch(operator_choice) {
         case 1:
             result = operand1 + operand2;
@@ -29,5 +31,5 @@ void main() {
             return;
     }
 
-    printf("Your result is: %d\n", result);
+    printf("Your result is: %g\n", result);
 }
